% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_preds_and_encs.R
\name{calc_enc_probs}
\alias{calc_enc_probs}
\title{Probability that prey encounters a predator in each cell}
\usage{
calc_enc_probs(df, grid_size = 15, reaction_dis = 0.5)
}
\arguments{
\item{df}{dataframe of predator counts per cell; use output from create_stream_raster_frame function}

\item{grid_size}{length of side of raster grid in meters; default is 15}

\item{reaction_dis}{maximum distance (in m) away from a predator that can trigger an encounter; default is 0.50}
}
\value{
a dataframe of encounter probabilities per grid cell for the entire reach
}
\description{
based on the idea that each predator will engage prey within a certain radius
total area occupied by predators divided by cell area is the encounter probability
}
\examples{
pred_pos <- get_pred_positions()
raster <- create_stream_raster_frame(pred_pos)
calc_enc_probs(raster)
}
