% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prey_survival_simulation.R
\name{calculate_final_status}
\alias{calculate_final_status}
\title{Final Status of Prey Fish}
\usage{
calculate_final_status(fish_frame, path_length)
}
\arguments{
\item{fish_frame}{dataframe with prey fish, lengths, survival modifiers, and simulation results; see simulate_encounters()}

\item{path_length}{number of cells traversed; determined by calculate_path_length()}
}
\value{
dataframe with a "final_status" column to indicate overall survival for each fish
}
\description{
Determines whether a fish survived the environment. All the outcomes (1 or 0) of encounters are summed. If the total is less than
the number of transects (i.e., a death (that is, a 0) occurred in a cell), then the fish is considered to have died.
}
