% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prey_survival_simulation.R
\name{calculate_num_cells_traversed}
\alias{calculate_num_cells_traversed}
\title{Number of Cells Traversed}
\usage{
calculate_num_cells_traversed(
  transect_length = 1000,
  n_transects = 20,
  grid_size = 15
)
}
\arguments{
\item{transect_length}{length of each transect in meters; default is 1000}

\item{n_transects}{integer of transects in the model; default is 20}

\item{grid_size}{length of side of raster grid in meters; default is 15}
}
\value{
integer value of cells traversed
}
\description{
Calculates the number of grid cells each fish traverses in the environment.
Assumes fish take the shortest path possible through the environment (i.e., in a straight line from left to right)
}
\examples{
calculate_num_cells_traversed(n_transects = 5)
calculate_num_cells_traversed(n_transects = 20)
}
