% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prey_survival_simulation.R
\name{calculate_proportion_of_survivors}
\alias{calculate_proportion_of_survivors}
\title{Calculate Proportion of Surviving fish}
\usage{
calculate_proportion_of_survivors(number_of_survivors, number_of_fish)
}
\arguments{
\item{number_of_survivors}{an integer value of surviving fish}

\item{number_of_fish}{the number of fish at the start of the model}
}
\value{
a proportion
}
\description{
Simply calculates the proportion of survivors relative to the number of fish at the start of the model
}
\examples{
calculate_proportion_of_survivors(number_of_survivors = 10, number_of_fish = 50)

}
