% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prey_survival_simulation.R
\name{calculate_total_survivors}
\alias{calculate_total_survivors}
\title{Calculate Total Survivors}
\usage{
calculate_total_survivors(final_status_frame)
}
\arguments{
\item{final_status_frame}{dataframe with final status of each fish; see calculate_final_status()}
}
\value{
an integer of the number of survivng prey fish
}
\description{
Simply calculates the total number of surviving fish based on the output of calculate_final_status()
}
