% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prey_survival_simulation.R
\name{combine_encounter_frame_and_fish_frame}
\alias{combine_encounter_frame_and_fish_frame}
\title{Joined Table of Prey Fish, Survival Length Modifiers, and Path}
\usage{
combine_encounter_frame_and_fish_frame(fish_frame, enc_frame)
}
\arguments{
\item{fish_frame}{a dataframe of fish, their lengths, and length survival boosts; see fish_with_length_survival_boost()}

\item{enc_frame}{a dataframe of fish and the encounter probabilities for predator in each cell on their path; see frame_of_all_cells_traversed_per_fish()}
}
\value{
a joined dataframe
}
\description{
joins the outputs from frame_of_all_cells_traversed_per_fish() and fish_with_length_survival_boost()
}
