% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prey_survival_simulation.R
\name{fish_with_length_survival_boost}
\alias{fish_with_length_survival_boost}
\title{Survival Boost from Length}
\usage{
fish_with_length_survival_boost(fish_frame)
}
\arguments{
\item{fish_frame}{dataframe of fish and their lengths; use output from the prey_fish_lengths function}
}
\value{
a dataframe with fish lengths and the effect of those lengths on predator encounters
}
\description{
Checks the effect a fish's length has on the probability of encountering a predator
}
\note{
this function can be parallelized; e.g., by setting plan(multisession)
}
\examples{
fish_frame <- prey_fish_lengths(number_of_fish = 10)
fish_with_length_survival_boost(fish_frame)
}
