% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prey_survival_simulation.R
\name{frame_of_all_cells_traversed_per_fish}
\alias{frame_of_all_cells_traversed_per_fish}
\title{Dataframe of Cells Traversed for Each Fish}
\usage{
frame_of_all_cells_traversed_per_fish(
  number_of_fish,
  enc_prob_vector,
  num_trans_traversed
)
}
\arguments{
\item{number_of_fish}{number (pos integer) of prey fish desired; should be the same as that for the prey_fish_lengths() function}

\item{enc_prob_vector}{vector of all encounter probabilities calculated for each cell; see calc_enc_probs()}

\item{num_trans_traversed}{number of cells traversed; determined by calculate_num_cells_traversed()}
}
\value{
a dataframe of encounter probabilities for each fish
}
\description{
Creates a dataframe of "paths" for each fish to "travel". A number of cells equal to num_trans_traversed
is randomly selected for each fish and becomes that fish's travel path.
}
\note{
this function can be parallelized; e.g., by setting plan(multisession)
}
\examples{
pred_pos <- get_pred_positions()
raster <- create_stream_raster_frame(pred_pos)
enc_prob_frame <- calc_enc_probs(raster)
enc_prob_vector <- enc_prob_frame \%>\% dplyr::pull(enc_prob)
num_trans = calculate_num_cells_traversed()
frame_of_all_cells_traversed_per_fish(number_of_fish = 10,
                                      enc_prob_vector = enc_prob_vector,
                                      num_trans_traversed = num_trans)


}
