% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_preds_and_encs.R
\name{get_pred_positions}
\alias{get_pred_positions}
\title{Predator positions across all transects}
\source{
defaults based on Michel et al. 2018. "Non-native fish predator density and molecular-based diet estimates suggest differing effects of predator species on Juvenile Salmon in the San Joaquin River, California"
}
\usage{
get_pred_positions(
  transect_length = 1000,
  n_transects = 20,
  lit_zone_size = 5,
  channel_width = 100
)
}
\arguments{
\item{transect_length}{length of each transect in meters; default is 1000}

\item{n_transects}{number of transects in the model; default is 20}

\item{lit_zone_size}{the size of the littoral zone (i.e., nearshore area) in meters; default is 5}

\item{channel_width}{width of the channel in meters; default is 100}
}
\value{
a dataframe of x, y positions for all predators
}
\description{
Calculates the position of all predators in terms of distance from shore and distance downstream for all transects
}
\examples{
get_pred_positions(n_transects = 5)
}
