% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_preds_and_encs.R
\name{graph_pred_positions}
\alias{graph_pred_positions}
\title{Visualization of Predator Positions}
\usage{
graph_pred_positions(df)
}
\arguments{
\item{df}{dataframe of predator positions; use output of get_pred_positions function}
}
\value{
a graph of all predator positions
}
\description{
graphs all predator positions in the 2D stream
}
\examples{
pred_pos <- get_pred_positions(n_transects = 5)
graph_pred_positions(pred_pos)

}
