% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_preds_and_encs.R
\name{lmb_calc}
\alias{lmb_calc}
\title{Numbers of largemouth bass in a reach}
\source{
defaults based on Michel et al. 2018. "Non-native fish predator density and molecular-based diet estimates suggest differing effects of predator species on Juvenile Salmon in the San Joaquin River, California"
}
\usage{
lmb_calc(n_transects, mean = 333, sd = 195)
}
\arguments{
\item{n_transects}{number of transects in the model}

\item{mean}{mean number of fish per transect; default is 333}

\item{sd}{std dev of number of fish per transect; default is 195}
}
\value{
a vector of numbers of fish per transect
}
\description{
Calculate the numbers of largemouth bass in a reach based on field data
}
\examples{
lmb_calc(n_transects = 10)
}
