% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prey_survival_simulation.R
\name{prey_fish_lengths}
\alias{prey_fish_lengths}
\title{Generate Prey Fish}
\source{
defaults based on Steel et al. 2020. "Applying the mean free-path length model to juvenile Chinook salmon migrating in the Sacramento River, California"
}
\usage{
prey_fish_lengths(
  number_of_fish,
  mean = 14,
  sd = (1.7/14) * mean,
  precision = 1
)
}
\arguments{
\item{number_of_fish}{number (pos integer) of prey fish desired}

\item{mean}{mean length of fish in cm; default is 14}

\item{sd}{std dev of fish length in cm; default is 1.7 and scales with mean}

\item{precision}{integer for level of precision for length (e.g., to the 10's or 100's place)}
}
\value{
dataframe of fish, each with a unique number and length
}
\description{
creates a dataframe of prey fish at a given mean size
}
\examples{
prey_fish_lengths(number_of_fish = 100)
}
