% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prey_survival_simulation.R
\name{simulate_encounters}
\alias{simulate_encounters}
\title{Encounter Simulations}
\usage{
simulate_encounters(combined_frame)
}
\arguments{
\item{combined_frame}{dataframe of prey fish, their lengths, survival boosts, and paths; see ()}
}
\value{
dataframe with all prior fish data plus columns indicating whether an encounter occurred and whether the fish survived
}
\description{
Simulates encounters for each fish in each cell. First checks whether an encounter occurs
based on fish length. If no encounter occurs, the fish survives. If an encounter occurs,
them a survival check occurs based on the abilities of the predator.
}
\note{
this function can be parallelized; e.g., by setting plan(multisession)
}
