% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_preds_and_encs.R
\name{striper_calc}
\alias{striper_calc}
\title{Numbers of striped bass in a reach}
\source{
defaults based on Michel et al. 2018. "Non-native fish predator density and molecular-based diet estimates suggest differing effects of predator species on Juvenile Salmon in the San Joaquin River, California"
}
\usage{
striper_calc(
  n_transects,
  agg_ratio = 1/3,
  mean_low = 60,
  sd_low = 43,
  mean_high = 870,
  sd_high = 491.5
)
}
\arguments{
\item{n_transects}{number of transects in the model}

\item{agg_ratio}{the ratio of large aggregations to low numbers of fish in a reach; default is 1/3}

\item{mean_low}{mean number of fish per transect when no aggregation is present; default is 60}

\item{sd_low}{std dev of number of fish per transect when no aggregation is present; default is 43}

\item{mean_high}{mean number of fish per transect when an aggregation is present; default is 870}

\item{sd_high}{std dev of number of fish per transect when an aggregation is present; default is 491.5}
}
\value{
a vector of numbers of fish per transect
}
\description{
Calculate the numbers of striped bass in a reach based on field data
}
\examples{
striper_calc(n_transects = 10)
}
