% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prey_survival_simulation.R
\name{survival_simulation_driver}
\alias{survival_simulation_driver}
\title{Run a Full Simulation}
\source{
defaults based on Steel et al. 2020. "Applying the mean free-path length model to juvenile Chinook salmon migrating in the Sacramento River, California"
and Michel et al. 2018. "Non-native fish predator density and molecular-based diet estimates suggest differing effects of predator species on Juvenile Salmon in the San Joaquin River, California"
}
\usage{
survival_simulation_driver(
  number_of_fish,
  mean_length = 14,
  sd_length = (1.7/14) * mean_length,
  transect_length = 1000,
  n_transects = 20,
  lit_zone_size = 5,
  channel_width = 100,
  grid_size = 15,
  reaction_dis = 0.5
)
}
\arguments{
\item{number_of_fish}{number (pos integer) of prey fish desired}

\item{mean_length}{mean length of fish in cm; default is 14}

\item{sd_length}{std dev of fish length in cm; default is 1.7 and scales with mean}

\item{transect_length}{length of each transect in meters; default is 1000}

\item{n_transects}{integer of transects in the model; default is 20}

\item{lit_zone_size}{the size of the littoral zone (i.e., nearshore area) in meters; default is 5}

\item{channel_width}{width of the channel in meters; default is 100}

\item{grid_size}{length of side of raster grid in meters; default is 15}

\item{reaction_dis}{maximum distance (in m) away from a predator that can trigger an encounter; default is 0.50}
}
\value{
the proportion of surviving fish
}
\description{
Runs a full simulation with a user-specified number of fish. Users can also adjust fish
mean length and sd, environment size, grid size, and predator reaction distance.
The model runs through the following:
calculates predators and their positions,
calculates grid cells and their encounter probabilities,
calculates a unique path for each fish,
simulates and resolves encounters for each fish in each cell,
determines survival after all fish have gone through each cell.
}
\details{
The return value is the proportion of survivors.
}
\note{
this function can be parallelized; e.g., by setting plan(multisession)
}
\examples{
survival_simulation_driver (number_of_fish = 20)
}
